requirements.txt
python-dotenv==1.0.1
schedule==1.2.2
vk_api==11.9.9
telebot
os
logging